+++ src/applications/model/udp-server.h
@@ -63,7 +63,7 @@ public:
    * \brief Returns the number of received packets
    * \return the number of received packets
    */
-  uint32_t GetReceived (void) const;
+  uint64_t GetReceived (void) const;
 
   /**
    * \brief Returns the size of the window used for checking loss.
@@ -98,7 +98,7 @@ private:
   uint16_t m_port; //!< Port on which we listen for incoming packets.
   Ptr<Socket> m_socket; //!< IPv4 Socket
   Ptr<Socket> m_socket6; //!< IPv6 Socket
-  uint32_t m_received; //!< Number of received packets
+  uint64_t m_received; //!< Number of received packets
   PacketLossCounter m_lossCounter; //!< Lost packet counter
 };
